requirements.txt
streamlit
langchain
langchain-community
google-generativeai
faiss-cpu
pymupdf
python-dotenv
sentence-transformers
